Zenith/JaiShreeRam/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
langchain==0.2.14
langchain-community==0.2.12
langchain-ollama==0.1.1
langchain-core==0.2.33
langgraph==0.2.4
requests==2.32.3
pyyaml==6.0.2
pydantic==2.8.2
google-generativeai>=0.3.0
langchain-google-genai>=0.0.3
chromadb>=0.5.0
sentence-transformers>=2.7.0
langchain-huggingface>=0.0.3
tree-sitter
tree-sitter-python
tree-sitter-javascript
tree-sitter-typescript
reportlab
GitPython